Made game() in script.js and updated playRound()
game() now has the functionality it is described to have.
Used an array to keep track of results for the end of the 5 rounds.
Used str.slice() to decode first character of playRound() string.
This character determines if it was a tie (0), win (1), or loss (2).
Used slice to console.log() the message without the first character.
Updated playRound() to switch-case based on player input,
rather than on computer input. This was necessary to correctly allow the
default case to occur when player gives faulty input.